requirements.txt
facenet-pytorch>=2.5
Pillow>=9.0
tqdm>=4.60
pandas>=1.5
scikit-learn>=1.2
matplotlib>=3.6
seaborn>=0.12
tabulate>=0.9
timm>=0.9.0
